+++ Common/src/main/java/de/markusbordihn/worlddimensionnexus/server/commands/TeleportCommand.java
@@ -19,22 +19,30 @@
 
 package de.markusbordihn.worlddimensionnexus.server.commands;
 
-import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.brigadier.builder.ArgumentBuilder;
 import com.mojang.brigadier.context.CommandContext;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import de.markusbordihn.worlddimensionnexus.commands.Command;
 import de.markusbordihn.worlddimensionnexus.server.commands.suggestions.DimensionSuggestion;
 import de.markusbordihn.worlddimensionnexus.teleport.TeleportCooldownManager;
+import de.markusbordihn.worlddimensionnexus.teleport.TeleportHistory;
 import de.markusbordihn.worlddimensionnexus.teleport.TeleportManager;
 import net.minecraft.ChatFormatting;
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.commands.Commands;
 import net.minecraft.commands.arguments.EntityArgument;
+import net.minecraft.commands.arguments.ResourceLocationArgument;
+import net.minecraft.core.registries.Registries;
 import net.minecraft.network.chat.Component;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.level.Level;
 
 public class TeleportCommand extends Command {
+
+  private static final String PLAYER_ARGUMENT = "player";
+  private static final String TELEPORTED_MESSAGE = "Teleported ";
 
   private TeleportCommand() {}
 
@@ -44,38 +52,56 @@ public class TeleportCommand extends Command {
             Commands.literal("dimension")
                 .requires(cs -> cs.hasPermission(Commands.LEVEL_MODERATORS))
                 .then(
-                    Commands.argument("name", StringArgumentType.word())
+                    Commands.argument("name", ResourceLocationArgument.id())
                         .suggests(DimensionSuggestion.DIMENSION_NAMES)
                         .executes(TeleportCommand::teleportToDimension)
                         .then(
-                            Commands.argument("player", EntityArgument.player())
+                            Commands.argument(PLAYER_ARGUMENT, EntityArgument.player())
                                 .executes(TeleportCommand::teleportPlayerToDimension))))
         .then(
             Commands.literal("back")
                 .executes(TeleportCommand::teleportBack)
                 .then(
-                    Commands.argument("player", EntityArgument.player())
+                    Commands.argument(PLAYER_ARGUMENT, EntityArgument.player())
                         .requires(cs -> cs.hasPermission(Commands.LEVEL_MODERATORS))
-                        .executes(TeleportCommand::teleportPlayerBack)));
+                        .executes(TeleportCommand::teleportPlayerBack)))
+        .then(
+            Commands.literal("overworld")
+                .executes(TeleportCommand::teleportToOverworld)
+                .then(
+                    Commands.argument(PLAYER_ARGUMENT, EntityArgument.player())
+                        .requires(cs -> cs.hasPermission(Commands.LEVEL_MODERATORS))
+                        .executes(TeleportCommand::teleportPlayerToOverworld)))
+        .then(
+            Commands.literal("history")
+                .executes(TeleportCommand::showTeleportHistory)
+                .then(
+                    Commands.argument(PLAYER_ARGUMENT, EntityArgument.player())
+                        .requires(cs -> cs.hasPermission(Commands.LEVEL_MODERATORS))
+                        .executes(TeleportCommand::showPlayerTeleportHistory)));
   }
 
   private static int teleportToDimension(final CommandContext<CommandSourceStack> context)
       throws CommandSyntaxException {
-    String dimensionName = StringArgumentType.getString(context, "name");
+    ResourceLocation dimensionLocation = ResourceLocationArgument.getId(context, "name");
     ServerPlayer player = context.getSource().getPlayerOrException();
 
-    if (TeleportManager.safeTeleportToDimension(player, dimensionName)) {
+    ResourceKey<Level> dimensionKey = ResourceKey.create(Registries.DIMENSION, dimensionLocation);
+    if (TeleportManager.safeTeleportToDimension(player, dimensionKey)) {
       return sendSuccessMessage(
           context.getSource(),
           Component.literal("Teleported to dimension ")
               .withStyle(ChatFormatting.GREEN)
-              .append(Component.literal(dimensionName).withStyle(ChatFormatting.YELLOW)));
+              .append(
+                  Component.literal(dimensionLocation.toString())
+                      .withStyle(ChatFormatting.YELLOW)));
     }
     return sendFailureMessage(
         context.getSource(),
         Component.literal("Failed to teleport to dimension ")
             .withStyle(ChatFormatting.RED)
-            .append(Component.literal(dimensionName).withStyle(ChatFormatting.YELLOW))
+            .append(
+                Component.literal(dimensionLocation.toString()).withStyle(ChatFormatting.YELLOW))
             .append(
                 Component.literal(". Dimension may not exist or be loaded.")
                     .withStyle(ChatFormatting.GRAY)));
@@ -83,19 +109,21 @@ public class TeleportCommand extends Command {
 
   private static int teleportPlayerToDimension(final CommandContext<CommandSourceStack> context)
       throws CommandSyntaxException {
-    String dimensionName = StringArgumentType.getString(context, "name");
-    ServerPlayer targetPlayer = EntityArgument.getPlayer(context, "player");
+    ResourceLocation dimensionLocation = ResourceLocationArgument.getId(context, "name");
+    ServerPlayer targetPlayer = EntityArgument.getPlayer(context, PLAYER_ARGUMENT);
 
-    if (TeleportManager.safeTeleportToDimension(targetPlayer, dimensionName)) {
+    ResourceKey<Level> dimensionKey = ResourceKey.create(Registries.DIMENSION, dimensionLocation);
+    if (TeleportManager.safeTeleportToDimension(targetPlayer, dimensionKey)) {
       return sendSuccessMessage(
           context.getSource(),
-          Component.literal("Teleported ")
+          Component.literal(TELEPORTED_MESSAGE)
               .withStyle(ChatFormatting.GREEN)
               .append(
                   Component.literal(targetPlayer.getName().getString())
                       .withStyle(ChatFormatting.YELLOW))
               .append(Component.literal(" to dimension ").withStyle(ChatFormatting.GREEN))
-              .append(Component.literal(dimensionName).withStyle(ChatFormatting.AQUA)));
+              .append(
+                  Component.literal(dimensionLocation.toString()).withStyle(ChatFormatting.AQUA)));
     }
     return sendFailureMessage(
         context.getSource(),
@@ -105,7 +133,8 @@ public class TeleportCommand extends Command {
                 Component.literal(targetPlayer.getName().getString())
                     .withStyle(ChatFormatting.YELLOW))
             .append(Component.literal(" to dimension ").withStyle(ChatFormatting.RED))
-            .append(Component.literal(dimensionName).withStyle(ChatFormatting.YELLOW)));
+            .append(
+                Component.literal(dimensionLocation.toString()).withStyle(ChatFormatting.YELLOW)));
   }
 
   private static int teleportBack(final CommandContext<CommandSourceStack> context)
@@ -153,7 +182,7 @@ public class TeleportCommand extends Command {
     if (TeleportManager.teleportBack(targetPlayer)) {
       return sendSuccessMessage(
           context.getSource(),
-          Component.literal("Teleported ")
+          Component.literal(TELEPORTED_MESSAGE)
               .withStyle(ChatFormatting.GREEN)
               .append(
                   Component.literal(targetPlayer.getName().getString())
@@ -169,5 +198,67 @@ public class TeleportCommand extends Command {
             .append(
                 Component.literal(targetPlayer.getName().getString())
                     .withStyle(ChatFormatting.YELLOW)));
+  }
+
+  private static int teleportToOverworld(final CommandContext<CommandSourceStack> context)
+      throws CommandSyntaxException {
+    ServerPlayer player = context.getSource().getPlayerOrException();
+
+    if (TeleportManager.teleportToDimensionWithoutHistory(player, Level.OVERWORLD)) {
+      return sendSuccessMessage(
+          context.getSource(),
+          Component.literal("Teleported to the Overworld").withStyle(ChatFormatting.GREEN));
+    }
+
+    return sendFailureMessage(
+        context.getSource(),
+        Component.literal("Failed to teleport to the Overworld").withStyle(ChatFormatting.RED));
+  }
+
+  private static int teleportPlayerToOverworld(final CommandContext<CommandSourceStack> context)
+      throws CommandSyntaxException {
+    ServerPlayer targetPlayer = EntityArgument.getPlayer(context, "player");
+
+    if (TeleportManager.teleportToDimensionWithoutHistory(targetPlayer, Level.OVERWORLD)) {
+      return sendSuccessMessage(
+          context.getSource(),
+          Component.literal(TELEPORTED_MESSAGE)
+              .withStyle(ChatFormatting.GREEN)
+              .append(
+                  Component.literal(targetPlayer.getName().getString())
+                      .withStyle(ChatFormatting.YELLOW))
+              .append(Component.literal(" to the Overworld").withStyle(ChatFormatting.AQUA)));
+    }
+
+    return sendFailureMessage(
+        context.getSource(),
+        Component.literal("Failed to teleport ")
+            .withStyle(ChatFormatting.RED)
+            .append(
+                Component.literal(targetPlayer.getName().getString())
+                    .withStyle(ChatFormatting.YELLOW))
+            .append(Component.literal(" to the Overworld").withStyle(ChatFormatting.RED)));
+  }
+
+  private static int showTeleportHistory(final CommandContext<CommandSourceStack> context)
+      throws CommandSyntaxException {
+    ServerPlayer player = context.getSource().getPlayerOrException();
+    String historyText = TeleportHistory.getFormattedPlayerHistory(player.getUUID());
+
+    Component historyComponent = Component.literal(historyText).withStyle(ChatFormatting.AQUA);
+    return sendSuccessMessage(context.getSource(), historyComponent);
+  }
+
+  private static int showPlayerTeleportHistory(final CommandContext<CommandSourceStack> context)
+      throws CommandSyntaxException {
+    ServerPlayer targetPlayer = EntityArgument.getPlayer(context, "player");
+    String historyText = TeleportHistory.getFormattedPlayerHistory(targetPlayer.getUUID());
+
+    Component historyComponent =
+        Component.literal(targetPlayer.getName().getString() + "'s ")
+            .withStyle(ChatFormatting.YELLOW)
+            .append(Component.literal(historyText).withStyle(ChatFormatting.AQUA));
+
+    return sendSuccessMessage(context.getSource(), historyComponent);
   }
 }