+++ CHANGELOG.md
@@ -2,9 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+# v0.2.0
+
+- Fixed and improved wiki documentation.
+- Added water dimension type.
+- Added fishing_dimension.wdn example file.
+- Improved teleport command and history features.
+
 # v0.1.0
 
 First beta release for Minecraft 1.21.1 with NeoForge for testing and feedback.
 
-⚠️ Beta Version Notice: This mod is currently in beta. 
+⚠️ Beta Version Notice: This mod is currently in beta.
 Features and commands may change before the stable release.